box: bind to free cpu at start
When configuring the database, create a shared memory segment, in which
store the information about the current instance:
* cpu id of the main thread
* checkpoint daemon schedule
If the segment already exists, read values of other instances in the
segment, and adjust own settings accordingly.
The segment is identifier by cluster UUID, but the segment name can be
reset in box.cfg{}.
The entire behaviour can be switched off in box.cfg.

closes Bind to the right CPU at start #1